Created the Login Model
The Login Model will handle all of the logic for identifying who the user
is. All you have to do is call, loginModel->getLoggedInUser() to get the
current logged in user.
Each model will have their own subfolder, and also has a list of model
components that separates the concerns. For example, the data access object,
the session manager, and the login validator.
Right now, all that is left to be implemented for the login model is the
Data access layer, as well as remember-me functionality.
Also, the database is working fine on my local machine.
To change the connection settings, go to
/web/application/config/database.php. There are instructions in the file.